+++ startup-dhcp.sh
@@ -28,7 +28,9 @@ do
     exit 1
   fi
 done
+
 PGPASSWORD=$DHCP_DBPASS /bin/sh -c "/usr/bin/psql -h $DHCP_DBHOST -U $DHCP_DBUSER -d $DHCP_DBNAME -a -f dhcpdb_create.sql"
+
 
 mkdir -p /usr/local/kea
 #  since the username and password is dyanmic with the k8s postgres operator
@@ -39,4 +41,6 @@ sed -i 's/EOF//g' /usr/local/kea/cray-dhcp-kea-dhcp4.conf
 #  helpful for debugging
 cat /usr/local/kea/cray-dhcp-kea-dhcp4.conf
 # what we use to run Cray DHCP Kea server
-/usr/local/sbin/kea-dhcp4 -c /usr/local/kea/cray-dhcp-kea-dhcp4.conf
+nohup /usr/local/sbin/kea-dhcp4 -c /usr/local/kea/cray-dhcp-kea-dhcp4.conf &
+# we will need to tune this
+while true; do /dhcp-helper.py; sleep 60; done